Fix the event type taking 4 bytes (instead of one) on the network.
It is also the occasion for a little refactoring.

+++ src/common/net/Deserializer.cc
@@ -4,7 +4,7 @@ namespace common {
 namespace net {
 
 EventId Deserializer::GetNextEventId(QDataStream &stream) {
-  quint32 id;
+  quint8 id;
   stream >> id;
   switch(id) {
     case EventId::kBombEventId:
